requirements.txt
asgiref==3.10.0
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
Django==5.2.7
djangorestframework==3.16.1
idna==3.11
isodate==0.7.2
lxml==6.0.2
platformdirs==4.5.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
requests-file==3.0.1
requests-toolbelt==1.0.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
xmlsec==1.3.16
zeep==4.3.2
